+++ API/src/main/java/fr/maxlego08/essentials/api/storage/IStorage.java
@@ -149,4 +149,6 @@ public interface IStorage {
     void upsertFlySeconds(UUID uniqueId, long flySeconds);
 
     long getFlySeconds(UUID uniqueId);
+
+    void deleteWorldData(String worldName);
 }