+++ src/main/java/fr/maxlego08/essentials/ZEssentialsPlugin.java
@@ -477,7 +477,7 @@ public final class ZEssentialsPlugin extends ZPlugin implements EssentialsPlugin
 
     @Override
     public void debug(String string) {
-        if (this.configuration.isEnableDebug()) {
+        if (this.configuration != null && this.configuration.isEnableDebug()) {
             this.getLogger().info(string);
         }
     }