+++ src/main/java/fr/maxlego08/essentials/commands/commands/items/CommandItemName.java
@@ -9,16 +9,44 @@ import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
 import fr.maxlego08.essentials.zutils.utils.paper.PaperComponent;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextDecoration;
+import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
+import org.bukkit.NamespacedKey;
+import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
+import org.bukkit.persistence.PersistentDataType;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class CommandItemName extends VCommand {
+
+    private final NamespacedKey itemNameRaw;
+
     public CommandItemName(EssentialsPlugin plugin) {
         super(plugin);
+
+        this.itemNameRaw = new NamespacedKey(plugin, "item-name-raw");
+
         this.setModule(ItemModule.class);
         this.setPermission(Permission.ESSENTIALS_ITEM_NAME);
         this.setDescription(Message.DESCRIPTION_ITEM_NAME);
-        this.addOptionalArg("name");
+        this.addOptionalArg("name", (sender, args) -> {
+            if (sender instanceof Player player) {
+                ItemStack itemStack = player.getInventory().getItemInMainHand();
+                if (itemStack.hasItemMeta()) {
+                    ItemMeta itemMeta = itemStack.getItemMeta();
+                    if (itemMeta.hasDisplayName()) {
+                        if (itemMeta.getPersistentDataContainer().has(this.itemNameRaw, PersistentDataType.STRING)) {
+                            return List.of(itemMeta.getPersistentDataContainer().getOrDefault(this.itemNameRaw, PersistentDataType.STRING, ""));
+                        }
+                        var component = itemMeta.displayName();
+                        return List.of(component == null ? "" : colorReverse(LegacyComponentSerializer.legacyAmpersand().serialize(component)));
+                    }
+                }
+            }
+            return new ArrayList<>();
+        });
         this.setExtendedArgs(true);
         this.onlyPlayers();
     }
@@ -42,6 +70,7 @@ public class CommandItemName extends VCommand {
             component = component.decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE);
         }
         itemMeta.displayName(component);
+        itemMeta.getPersistentDataContainer().set(this.itemNameRaw, PersistentDataType.STRING, itemName);
 
         itemStack.setItemMeta(itemMeta);