+++ src/main/java/fr/maxlego08/essentials/storage/database/repositeries/UserHomeRepository.java
@@ -29,8 +29,8 @@ public class UserHomeRepository extends Repository {
         return select(HomeDTO.class, schema -> schema.where("unique_id", uuid));
     }
 
-    public void deleteHomes(UUID uuid, String name) {
-        delete(table -> table.where("unique_id", name).where("name", name));
+    public void deleteHome(UUID uuid, String name) {
+        delete(table -> table.where("unique_id", uuid).where("name", name));
     }
 
     public List<Home> getHomes(UUID uuid, String homeName) {
@@ -39,5 +39,9 @@ public class UserHomeRepository extends Repository {
 
     public List<Home> getHomes(UUID uuid) {
         return select(HomeDTO.class, schema -> schema.where("unique_id", uuid)).stream().map(homeDTO -> (Home) new ZHome(stringAsLocation(homeDTO.location()), homeDTO.name(), null)).toList();
+    }
+
+    public void deleteWorldData(String worldName) {
+        delete(table -> table.where("location", "LIKE", "%" + worldName + "%"));
     }
 }