+++ src/main/java/fr/maxlego08/essentials/zutils/utils/component/adapters/TextAdapter.java
@@ -1,12 +1,12 @@
-package fr.maxlego08.essentials.hooks.protocollib.component.adapters;
+package fr.maxlego08.essentials.zutils.utils.component.adapters;
 
 import com.google.gson.JsonObject;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import fr.maxlego08.essentials.hooks.protocollib.component.AbstractComponent;
-import fr.maxlego08.essentials.hooks.protocollib.component.components.ClickEvent;
-import fr.maxlego08.essentials.hooks.protocollib.component.components.HoverEvent;
-import fr.maxlego08.essentials.hooks.protocollib.component.components.TextComponent;
+import fr.maxlego08.essentials.zutils.utils.component.AbstractComponent;
+import fr.maxlego08.essentials.zutils.utils.component.components.ClickEvent;
+import fr.maxlego08.essentials.zutils.utils.component.components.HoverEvent;
+import fr.maxlego08.essentials.zutils.utils.component.components.TextComponent;
 
 import java.io.IOException;